[cypress] Isolating login flow in a file
